.claude/templates/init-dev/stacks/python/requirements.txt
# Web Framework
Django>=4.2.0
# or
# FastAPI>=0.100.0
# uvicorn[standard]>=0.20.0

# Environment variables (.env support)
python-dotenv>=1.0.0

# Development dependencies
watchdog>=3.0.0

# Common utilities
requests>=2.31.0